'controllerUser'

+++ BidBackend/src/main/java/com/example/BidBackend/service/ContratDeVenteService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ContratDeVenteService {
 
-    ContratDeVente generateSaleContract(int articleId, int buyerId);
+    ContratDeVente generateSaleContract(long articleId, int buyerId);
     public ContratDeVente save(ContratDeVente contratDeVente);
 
     ContratDeVente findById(int id);
